# src/main/resources/openapi/ps_api.yaml
openapi: "3.0.0"
info:
  title: Felho lab photo album
  description: Felho lab photo album
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development
security:
  -  oidcAuth: []

tags:
  - name: Authentication
    description: User registration, login, and logout
  - name: Photos
    description: Photo management operations

paths:
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Create a new user account
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: User successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: Authenticate user and receive access token
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: Successfully logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      description: Invalidate the current user session
      operationId: logoutUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successfully logged out
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Successfully logged out
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /photos:
    get:
      tags:
        - Photos
      summary: List all photos
      description: Retrieve a list of all photos with names and upload dates, sorted by name or date
      operationId: listPhotos
      parameters:
        - name: sortBy
          in: query
          description: Sort photos by name or date
          required: false
          schema:
            type: string
            enum:
              - name
              - date
            default: date
        - name: order
          in: query
          description: Sort order (ascending or descending)
          required: false
          schema:
            type: string
            enum:
              - asc
              - desc
            default: desc
      responses:
        '200':
          description: List of photos retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  photos:
                    type: array
                    items:
                      $ref: '#/components/schemas/PhotoListItem'
                  total:
                    type: integer
                    description: Total number of photos
                    example: 25

    post:
      tags:
        - Photos
      summary: Upload a new photo
      description: Upload a new photo with a name (requires authentication)
      operationId: uploadPhoto
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - name
                - file
              properties:
                name:
                  type: string
                  maxLength: 40
                  description: Name of the photo (max 40 characters)
                  example: Summer vacation sunset
                file:
                  type: string
                  format: binary
                  description: The photo file to upload
      responses:
        '201':
          description: Photo uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Photo'
        '400':
          description: Invalid input (e.g., name too long, invalid file format)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - Login required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: File too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /photos/{photoId}:
    get:
      tags:
        - Photos
      summary: Get photo details
      description: Retrieve details and the image of a specific photo
      operationId: getPhoto
      parameters:
        - name: photoId
          in: path
          required: true
          description: Unique identifier of the photo
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Photo details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Photo'
        '404':
          description: Photo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Photos
      summary: Delete a photo
      description: Delete a specific photo (requires authentication)
      operationId: deletePhoto
      security:
        - bearerAuth: []
      parameters:
        - name: photoId
          in: path
          required: true
          description: Unique identifier of the photo
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Photo deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Photo deleted successfully
        '401':
          description: Unauthorized - Login required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - User does not own this photo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Photo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /photos/{photoId}/image:
    get:
      tags:
        - Photos
      summary: Get photo image
      description: Retrieve the actual image file of a photo
      operationId: getPhotoImage
      parameters:
        - name: photoId
          in: path
          required: true
          description: Unique identifier of the photo
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Photo image retrieved successfully
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
        '404':
          description: Photo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint

  schemas:
    User:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
          format: password

    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique user identifier
          example: 550e8400-e29b-41d4-a716-446655440000
        username:
          type: string
          description: User's username
          example: johndoe
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
          example: '2026-02-16T10:30:00Z'

    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT access token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        expiresIn:
          type: integer
          description: Token expiration time in seconds
          example: 3600
        user:
          $ref: '#/components/schemas/UserResponse'

    PhotoListItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique photo identifier
          example: 1
        name:
          type: string
          maxLength: 40
          description: Name of the photo
          example: Summer vacation sunset
        uploadDate:
          type: string
          format: date-time
          description: Date and time when the photo was uploaded (year-month-day hour:minute)
          example: '2026-02-16T14:30:00Z'

    Photo:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique photo identifier
          example: 1
        name:
          type: string
          maxLength: 40
          description: Name of the photo
          example: Summer vacation sunset
        uploadDate:
          type: string
          format: date-time
          description: Date and time when the photo was uploaded (year-month-day hour:minute)
          example: '2026-02-16T14:30:00Z'
        imageUrl:
          type: string
          format: uri
          description: URL to retrieve the photo image
          example: /api/v1/photos/1/image
        uploadedBy:
          type: integer
          format: int64
          description: ID of the user who uploaded the photo
          example: 1
        mimeType:
          type: string
          description: MIME type of the photo
          example: image/jpeg

    Error:
      type: object
      properties:
        code:
          type: string
          description: Error code
          example: VALIDATION_ERROR
        message:
          type: string
          description: Human-readable error message
          example: Photo name must not exceed 40 characters
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                description: Field that caused the error
                example: name
              message:
                type: string
                description: Detailed error message for this field
                example: Name exceeds maximum length of 40 characters